backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
packaging==23.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
zmq==0.0.0
paho-mqtt
marshmallow
cachetools
gunicorn
psycopg2-binary